scripts/requirements.txt
black==22.1.0
click==8.0.3
colour @ git+https://github.com/colour-science/colour.git@efdadd5cc0e8131b3245420e2ae4582eaec4f668
imageio==2.14.1
mypy-extensions==0.4.3
numpy==1.22.1
pathspec==0.9.0
Pillow==9.0.0
platformdirs==2.4.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scipy==1.7.3
tomli==2.0.0
